requirements.txt
fastapi
pydantic
dynaconf
tortoise-orm

aerich==0.6.3
